Gracefully handle executor termination
Uses ZMQ_ROUTER_MANDATORY option so that impropertly
routed messages will be flagged as errors. Messages
intended to be sent to bad addresses get queued up
in the backlog, to be handled as soon as the next
executor is free.

Typing this out, I realize that the message should be
attempted to be sent to every known available executor
rather than just the first.